requirements/development.txt
-r base.txt
django-debug-toolbar>=1.8
Werkzeug>=0.12.2
astroid==1.6.1
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.2
pylint-django==0.9.1
pylint-plugin-utils==0.2.6
wrapt==1.10.11
